WIP on compress; need to test
This is probably a bad idea which is why it's being shelved.
Compression may help efficiency a bit, but the main use is in making a
term understandable by a human.  Simpler memoization of raw terms
without separating by lambda depth or dependency tiers would be fine for
this purpose.